refact(front): cambiando nombe de variables y estados en archivos useGet y usePost
